requirements.txt
langchain
langchain-together
python-dotenv
requests
pyttsx3